requirements.txt
PyYAML
# python==3.6.3
certifi==2018.1.18
Django==2.0
djangorestframework==3.7.7
pytz==2018.3
django-cors-headers
psycopg2
lxml
schedule
numpy
dbfread
# remove in production below
django-sslserver